+++ spec/integrations/swarm/forceful_nodes_restart_graceful_supervisor_stop_spec.rb
@@ -37,6 +37,19 @@ module Karafka
           end
         end
       end
+
+      def on_connection_listener_fetch_loop(_event)
+        periodically do
+          Kernel.exit!(orphaned_exit_code) if node.orphaned?
+
+          if DT[:forks].size > 2
+            node.healthy
+          else
+            # Fake hang it forever
+            sleep
+          end
+        end
+      end
     end
   end
 end